feat: connects connect data-sources api to ui
